backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
supabase>=2.3.0
python-dotenv>=1.0.0
pymupdf>=1.24.0
spacy>=3.7.0
scikit-learn>=1.5.0
networkx>=3.3
nltk>=3.8.1
yt-dlp>=2024.5.0
youtube-transcript-api>=0.6.2
requests>=2.31.0
beautifulsoup4>=4.12.3
numpy>=1.26.0
